Organize AL items with folders
