+++ ArchitectsLibrary/ArchitectsLibrary.csproj
@@ -192,27 +192,27 @@
     <Compile Include="Interfaces\IVehicleOnEquip.cs" />
     <Compile Include="Interfaces\IVehicleOnToggleRepeating.cs" />
     <Compile Include="Interfaces\ISeaMothOnUse.cs" />
-    <Compile Include="Items\AlienCompositeGlass.cs" />
+    <Compile Include="Items\AdvancedMaterials\AlienCompositeGlass.cs" />
+    <Compile Include="Items\AdvancedMaterials\CobaltIngot.cs" />
+    <Compile Include="Items\AdvancedMaterials\PrecursorAlloyIngot.cs" />
+    <Compile Include="Items\AdvancedMaterials\ReinforcedGlass.cs" />
     <Compile Include="Items\AotuPoster.cs" />
-    <Compile Include="Items\CobaltIngot.cs" />
-    <Compile Include="Items\Cobalt.cs" />
+    <Compile Include="Items\Cubes\Electricube.cs" />
+    <Compile Include="Items\Cubes\OmegaCube.cs" />
+    <Compile Include="Items\Cubes\PrecursorIonCube.cs" />
+    <Compile Include="Items\Cubes\RedIonCube.cs" />
     <Compile Include="Items\CustomBuilder.cs" />
-    <Compile Include="Items\DrillableCobalt.cs" />
-    <Compile Include="Items\DrillableRedBeryl.cs" />
-    <Compile Include="Items\DrillableMorganite.cs" />
-    <Compile Include="Items\DrillableSapphire.cs" />
-    <Compile Include="Items\DrillableEmerald.cs" />
-    <Compile Include="Items\Electricube.cs" />
-    <Compile Include="Items\Morganite.cs" />
-    <Compile Include="Items\OmegaCube.cs" />
-    <Compile Include="Items\PrecursorIonCube.cs" />
-    <Compile Include="Items\RedBeryl.cs" />
-    <Compile Include="Items\RedIonCube.cs" />
-    <Compile Include="Items\ReinforcedGlass.cs" />
-    <Compile Include="Items\Sapphire.cs" />
-    <Compile Include="Items\Emerald.cs" />
-    <Compile Include="Items\PrecursorAlloyIngot.cs" />
+    <Compile Include="Items\Drillables\DrillableCobalt.cs" />
+    <Compile Include="Items\Drillables\DrillableEmerald.cs" />
+    <Compile Include="Items\Drillables\DrillableMorganite.cs" />
+    <Compile Include="Items\Drillables\DrillableRedBeryl.cs" />
+    <Compile Include="Items\Drillables\DrillableSapphire.cs" />
+    <Compile Include="Items\Minerals\Cobalt.cs" />
+    <Compile Include="Items\Minerals\Emerald.cs" />
+    <Compile Include="Items\Minerals\Morganite.cs" />
+    <Compile Include="Items\Minerals\RedBeryl.cs" />
     <Compile Include="Buildables\PrecursorFabricator.cs" />
+    <Compile Include="Items\Minerals\Sapphire.cs" />
     <Compile Include="Items\VanillaPrefabPatching\PrecursorIonBattery.cs" />
     <Compile Include="Items\VanillaPrefabPatching\PrecursorIonCrystal.cs" />
     <Compile Include="Items\VanillaPrefabPatching\PrecursorIonPowerCell.cs" />
