+++ ArchitectsLibrary/Items/Cubes/RedIonCube.cs
@@ -1,4 +1,4 @@
-﻿namespace ArchitectsLibrary.Items
+﻿namespace ArchitectsLibrary.Items.Cubes
 {
     using SMLHelper.V2.Crafting;
     using System.Collections.Generic;
