+++ ArchitectsLibrary/Items/Drillables/DrillableSapphire.cs
@@ -1,4 +1,4 @@
-﻿namespace ArchitectsLibrary.Items
+﻿namespace ArchitectsLibrary.Items.Drillables
 {
     using UnityEngine;
     using API;
