+++ ArchitectsLibrary/Items/Minerals/Morganite.cs
@@ -1,4 +1,4 @@
-﻿namespace ArchitectsLibrary.Items
+﻿namespace ArchitectsLibrary.Items.Minerals
 {
     using System.Collections.Generic;
     using UnityEngine;
