+++ ArchitectsLibrary/Main.cs
@@ -7,6 +7,10 @@ using System.Collections.Generic;
 using System.Reflection;
 using ArchitectsLibrary.API;
 using ArchitectsLibrary.Items;
+using ArchitectsLibrary.Items.AdvancedMaterials;
+using ArchitectsLibrary.Items.Cubes;
+using ArchitectsLibrary.Items.Drillables;
+using ArchitectsLibrary.Items.Minerals;
 using ArchitectsLibrary.Items.VanillaPrefabPatching;
 using ArchitectsLibrary.Patches;
 using ArchitectsLibrary.Handlers;
